refactor: restructure pubmed module to match pmc organization
Break down the large pubmed.rs file into a well-organized module structure:

- src/pubmed/models.rs: Core domain models (PubMedArticle)
- src/pubmed/responses.rs: Internal API response structures
- src/pubmed/parser.rs: XML parsing logic (PubMedXmlParser)
- src/pubmed/client.rs: Client implementation and HTTP logic
- src/pubmed/mod.rs: Module declaration and public re-exports

This matches the existing src/pmc/ module organization and provides:
- Better separation of concerns
- Improved maintainability
- Consistent codebase structure
- Easier testing and future development

Tests updated to use PubMedXmlParser::parse_article_from_xml()
instead of the client method. All functionality preserved.
