+++ src/pubmed/mod.rs
@@ -1,0 +1,14 @@
+//! PubMed client for searching and fetching article metadata
+//!
+//! This module provides functionality to interact with PubMed E-utilities APIs
+//! for searching biomedical literature and retrieving article metadata.
+
+pub mod client;
+pub mod models;
+pub mod parser;
+pub mod responses;
+
+// Re-export public types
+pub use client::PubMedClient;
+pub use models::PubMedArticle;
+pub use parser::PubMedXmlParser;
