+++ src/pubmed/models.rs
@@ -1,0 +1,22 @@
+use serde::{Deserialize, Serialize};
+
+/// Represents a PubMed article with metadata
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct PubMedArticle {
+    /// PubMed ID
+    pub pmid: String,
+    /// Article title
+    pub title: String,
+    /// List of authors
+    pub authors: Vec<String>,
+    /// Journal name
+    pub journal: String,
+    /// Publication date
+    pub pub_date: String,
+    /// DOI (Digital Object Identifier)
+    pub doi: Option<String>,
+    /// Abstract text (if available)
+    pub abstract_text: Option<String>,
+    /// Article types (e.g., "Clinical Trial", "Review", etc.)
+    pub article_types: Vec<String>,
+}
